// admin-pwa/src/app/models/app-info.model.ts
export interface AppInfo {
  androidAppCurrentVersion: number;
  iosAppCurrentVersion: number;
  androidPackageName: string;
  iOsAppId: string;
  appEmail: string;
  privacyPolicyUrl: string;
  termsOfServicesUrl: string;
  subscriptionIds: string[];
  freeAccountMaxDistance: number;
  vipAccountMaxDistance: number;
  agoraAppID: string;
}